requirements.txt
astor==0.8.1
h5py==3.7.0
imageio==2.19.2
importlib-metadata==4.11.4
numpy==1.21.6
opencv-python==4.5.5.64
paddlepaddle==2.3.0
pandas==1.3.5
Pillow==9.1.1
protobuf==3.20.1
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
tensorlayerx==0.5.7
tensorflow==2.0.0
tqdm==4.64.1
urllib3==1.26.9
chardet
scikit-image
shapely